make public routes folder
