+++ client/src/components/PublicRoutes/Hero/Footer.js
